Rename SolverButton to FloatingSolverButton
