+++ packages/scrabble-solver/src/components/Solver/components/FloatingSolveButton/FloatingSolveButton.module.scss
@@ -1,6 +1,6 @@
 @import 'styles/animations';
 
-.solveButton {
+.floatingSolveButton {
   padding: var(--spacing--l);
   border-radius: 50%;
   box-shadow: var(--box-shadow) !important;
