+++ packages/scrabble-solver/src/components/Solver/components/FloatingSolveButton/FloatingSolveButton.tsx
@@ -16,14 +16,14 @@ import {
 import Button from '../../../Button';
 import Spinner from '../../../Spinner';
 
-import styles from './SolveButton.module.scss';
+import styles from './FloatingSolveButton.module.scss';
 
 interface Props {
   className?: string;
   onClick?: MouseEventHandler;
 }
 
-const SolveButton: FunctionComponent<Props> = ({ className, onClick = noop }) => {
+const FloatingSolveButton: FunctionComponent<Props> = ({ className, onClick = noop }) => {
   const dispatch = useDispatch();
   const translate = useTranslate();
   const isLoading = useTypedSelector(selectIsLoading);
@@ -39,7 +39,7 @@ const SolveButton: FunctionComponent<Props> = ({ className, onClick = noop }) =>
   return (
     <Button
       aria-label={translate('results.solve')}
-      className={classNames(styles.solveButton, className)}
+      className={classNames(styles.floatingSolveButton, className)}
       disabled={isLoading || !isOutdated || !hasTiles}
       Icon={isLoading ? Spinner : Search}
       iconClassName={classNames({
@@ -54,4 +54,4 @@ const SolveButton: FunctionComponent<Props> = ({ className, onClick = noop }) =>
   );
 };
 
-export default SolveButton;
+export default FloatingSolveButton;
