+++ packages/scrabble-solver/src/components/Solver/components/FloatingSolveButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FloatingSolveButton';
