--- packages/scrabble-solver/src/components/Solver/components/SolveButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SolveButton';
