+++ packages/scrabble-solver/src/components/Solver/components/index.ts
@@ -1,4 +1,4 @@
 export { default as EmptyState } from './EmptyState';
+export { default as FloatingSolveButton } from './FloatingSolveButton';
 export { default as InsertButton } from './InsertButton';
 export { default as ResultCandidatePicker } from './ResultCandidatePicker';
-export { default as SolveButton } from './SolveButton';
